projeto/requirements.txt
opencv-python
pillow
pyautogui
pygetwindow
